refactor, remove function to utils
